requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.38
ta>=0.10.2
streamlit>=1.28.0
plotly>=5.18.0
python-telegram-bot>=20.6
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0